Encapsulate Server methods into Server class
