+++ server/Dockerfile
@@ -1,3 +1,3 @@
 FROM python:3.8
-COPY server/server.py /server.py 
+COPY server/* /
 ENTRYPOINT ["/bin/sh"]
